orbital_backend/easysearchcrawl/requirements.txt
Flask==2.0.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
Pillow==7.2.0
prompt-toolkit==3.0.18
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.6
Pygments==2.9.0
pyparsing==2.4.7
Pyrebase-t==1.0.28
Pyrebase4==4.4.3
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
selenium==3.141.0
flask-cors==3.0.10